% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nyc_subway_routes_offset_sf}
\alias{nyc_subway_routes_offset_sf}
\title{NYC subway route shapes with schematic offsets}
\format{
\subsection{\code{nyc_subway_routes_offset_sf}}{

A simple feature collection with 311 rows and 7 columns:
\describe{
\item{shape_id}{GTFS shape identifier.}
\item{route_id}{GTFS route identifier.}
\item{route_short_name}{Short service name (bullet letter/number).}
\item{route_long_name}{Descriptive long name.}
\item{route_color}{Brand color as a hex string prefixed with \verb{#}.}
\item{x_offset}{Horizontal offset in feet applied to the geometry.}
\item{geometry}{Offset linestring geometry in EPSG:2263.}
}
}
}
\source{
\url{https://new.mta.info/developers}
}
\usage{
nyc_subway_routes_offset_sf
}
\description{
A version of \link{nyc_subway_routes_sf} in which each route has been
shifted horizontally by a small amount (in feet, CRS EPSG:2263)
so that co-running services can be drawn as parallel lines in
schematic maps instead of stacking on top of each other.
EPSG:2263, NAD83 / New York Long Island (ftUS).
}
\details{
Offsets are computed as \code{(group_id - n_routes / 2) * 50}, so each
route is shifted by a multiple of 50 feet relative to the middle
route. This is a cosmetic transformation intended only for
schematic visualization; for any spatial analysis use
\link{nyc_subway_routes_sf} instead.
}
\author{
Kieran Healy
}
\keyword{datasets}
