% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nyc_subway_stops_parent_sf}
\alias{nyc_subway_stops_parent_sf}
\title{NYC subway parent stations}
\format{
\subsection{\code{nyc_subway_stops_parent_sf}}{

A simple feature collection with 496 rows and 5 columns:
\describe{
\item{stop_id}{GTFS stop identifier for the parent station.}
\item{stop_name}{Station name.}
\item{location_type}{GTFS location type; always \code{1} here.}
\item{parent_station}{Always \code{NA} for parent stations.}
\item{geometry}{Point geometry in EPSG:2263.}
}
}
}
\source{
\url{https://new.mta.info/developers}
}
\usage{
nyc_subway_stops_parent_sf
}
\description{
Point locations of the 496 parent subway stations
(\code{location_type == 1}). Each parent station corresponds to one or
more directional platforms in \link{nyc_subway_stops_platform_sf}.
EPSG:2263, NAD83 / New York Long Island (ftUS).
}
\details{
Filtered from \link{nyc_subway_stops_sf} on \code{location_type == 1}.
}
\author{
Kieran Healy
}
\keyword{datasets}
