% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nyc_subway_stops_platform_sf}
\alias{nyc_subway_stops_platform_sf}
\title{NYC subway directional platforms}
\format{
\subsection{\code{nyc_subway_stops_platform_sf}}{

A simple feature collection with 992 rows and 5 columns:
\describe{
\item{stop_id}{GTFS platform identifier, typically the parent
\code{stop_id} with a trailing \code{N} or \code{S} for direction.}
\item{stop_name}{Station name (usually the same as the parent).}
\item{location_type}{GTFS location type; always \code{NA} here.}
\item{parent_station}{\code{stop_id} of the owning parent station.}
\item{geometry}{Point geometry in EPSG:2263.}
}
}
}
\source{
\url{https://new.mta.info/developers}
}
\usage{
nyc_subway_stops_platform_sf
}
\description{
Point locations of the 992 directional platforms (child stops
with \code{is.na(location_type)}). Each platform has a \code{parent_station}
id linking to a row in \link{nyc_subway_stops_parent_sf}. EPSG:2263,
NAD83 / New York Long Island (ftUS).
}
\details{
Filtered from \link{nyc_subway_stops_sf} on \code{is.na(location_type)}.
Directional platforms are the rows referenced by GTFS
\code{stop_times.txt} and are what should be joined against trip
data.
}
\author{
Kieran Healy
}
\keyword{datasets}
