% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nyc_subway_transfers_df}
\alias{nyc_subway_transfers_df}
\title{NYC subway transfers}
\format{
\subsection{\code{nyc_subway_transfers_df}}{

A tibble with 613 rows and 4 columns:
\describe{
\item{from_stop_id}{\code{stop_id} where the transfer originates.}
\item{to_stop_id}{\code{stop_id} where the transfer terminates.}
\item{transfer_type}{GTFS transfer type code. \code{0} = recommended
transfer, \code{2} = transfer requires a minimum time, \code{3} = no
transfer possible.}
\item{min_transfer_time}{Minimum time in seconds to complete
the transfer; \code{NA} if not specified.}
}
}
}
\source{
\url{https://new.mta.info/developers}
}
\usage{
nyc_subway_transfers_df
}
\description{
Transfer pairs between subway stops, from the GTFS
\code{transfers.txt} file.
}
\details{
Derived from \code{transfers.txt} in the MTA NYC Transit subway GTFS
feed. Stop ids can be joined to \link{nyc_subway_stops_sf}.
}
\author{
Kieran Healy
}
\keyword{datasets}
